examples/qa_with_sources/server/requirements.txt
steamship-langchain==0.0.10rc1